Improved handling for auth_basic authentication
Sometimes it's usefull to let the main webserver (e.g. apache) do all
the auth. So you can use all auth modules of the webserver e.g. ldap for
single sign on.

If the option 'trust_basic_auth' is enabled no password is checked. You
MUST(!!!) enable some auth in your webserver (e.g. htaccess etc).

Every account which exists in database is granted the access.

To create an account at first login the Option :create_auth_ondemand are
used if it is enabled a new account will be created on every successfull
webserver login without the expected user in masquerade database.

As an email address the login name and the domain configured in
'default_mail_domain' is used. A random password is generated because
some password is needed. This password is only for database checks but
should never be used. The Webserver handles all auth and password
changes are independend of masquerade. So the generated password is only
a dummy value.