added option for random password on automatic account creation
if create_auth_ondemand:random_password is enabled a random password is
generated - user can't login with this password because its secret.
This is usefull if auth is always over httpbasic and anything else is
disabled. We must set a password so wie generate a secret password.

Else the auth_basic password is copied to local storage. This can be
usefull if the user use the masquerade login sometimes. But take care of
password changes! http_basic and masquerade passwords are not stay in
sync.